requirements.txt
numpy==1.24.3
matplotlib==3.7.2
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
tensorflow==2.13.0
PyYAML==6.0.1
Jinja2==3.1.2